productivity-tracker-app/src/backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
chromadb
pydantic
numpy
langchain-community
tqdm
tiktoken
langchain-openai